Adding to pipeline
